Add updateDone signal to start notifyService each time updateService complete
